pyproject.toml
[project]
name = "dnd-module-gen"
version = "0.1.0"
description = "D&D module converter for FoundryVTT - converts official D&D PDFs into structured XML assets"
requires-python = ">=3.9"
dependencies = [
    "google-genai",
    "python-dotenv>=1.0.0",
    "PyMuPDF>=1.23.0",
    "Pillow>=10.0.0",
    "pytesseract>=0.3.10",
    "pytest>=8.4.2",
    "pytest-anyio>=0.0.0",
    "pytest-asyncio>=1.0.0",
    "pytest-rerunfailures>=14.0",
    "pytest-order>=1.3.0",
    "pytest-timeout>=2.0.0",
    "pytest-xdist>=3.0.0",
    "requests>=2.31.0",
    "aiohttp>=3.9.0",
    "opencv-python>=4.9.0",
    "scikit-image>=0.22.0",
    # UI Backend dependencies
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "websockets>=12.0",
    "pydantic>=2.0.0",
    "httpx>=0.25.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
